requirements.txt
# Core dependencies
openai>=1.6.0
numpy>=1.24.0,<2.0.0
python-dotenv>=1.0.0

# Optional but recommended for development
ipython>=8.12.0
pytest>=7.4.0
pytest-cov>=4.1.0

# For type hints (optional)
typing-extensions>=4.8.0

# For better logging (optional)
colorlog>=6.7.0

# For database migrations (optional)
alembic>=1.12.0
